# Memoize recurse regex